Move `output` -> `lib/output`
