+++ src/lib/output/get-module-output.js
