+++ src/lib/output/get-runtime-import.js
@@ -2,12 +2,12 @@ import path from 'path';
 
 import {stringifyRequest} from 'loader-utils';
 
-import {getImportStr} from '../utils/get-import-str';
+import {getImportStr} from '../../utils/get-import-str';
 
 export function getRuntimeImport(loaderContext, esModule) {
     const runtimePath = stringifyRequest(
         loaderContext,
-        `${path.resolve(path.join(__dirname, '..', 'public', 'runtime.js'))}`
+        `${path.resolve(path.join(__dirname, '..', '..', 'public', 'runtime.js'))}`
     );
 
     return `${getImportStr(runtimePath, esModule)('runtime')}`
